requirements.txt
astroid==2.11.7
autopep8==1.6.0
dill==0.3.5.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
pycodestyle==2.9.0
pylint==2.14.5
PyYAML==6.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
typing_extensions==4.3.0
wrapt==1.14.1
